add streaming unit tests
