+++ rows/fixtures/result2.jsonl
@@ -1,0 +1,3 @@
+{  "message_type": "START",  "query_id": "16FF2A0300ECA753",  "result_columns": [    {      "name": "int_col",      "type": "int null"    }  ]}
+{"message_type":"DATA",  "data": [    [      3    ],    [      null    ]  ]}
+{"message_type":"FINISHED_WITH_SUCCESS","statistics":{"elapsed":0.001797702,"rows_read":2,"bytes_read":293,"time_before_execution":0.001251613,"time_to_execute":0.000544098,"scanned_bytes_cache":2003,"scanned_bytes_storage":0}}
